move couch case rebuild to CouchProcessor
