+++ corehq/ex-submodules/casexml/apps/case/cleanup.py
@@ -40,26 +40,6 @@ def close_case(case_id, domain, user):
     return submit_case_blocks([case_block], domain, username, user_id)
 
 
-def _get_actions_from_forms(sorted_forms, case_id):
-    from corehq.apps.commtrack.processing import get_stock_actions
-    case_actions = []
-    domain = None
-    for form in sorted_forms:
-        if domain is None:
-            domain = form.domain
-        assert form.domain == domain
-
-        case_updates = get_case_updates(form)
-        filtered_updates = [u for u in case_updates if u.id == case_id]
-        for u in filtered_updates:
-            case_actions.extend(u.get_case_actions(form))
-        stock_actions = get_stock_actions(form)
-        case_actions.extend([intent.action
-                             for intent in stock_actions.case_action_intents
-                             if not intent.is_deprecation])
-    return case_actions, domain
-
-
 def rebuild_case_from_actions(case, actions):
     strategy = ActionsUpdateStrategy(case)
     strategy.reset_case_state()
@@ -80,35 +60,7 @@ def rebuild_case_from_forms(domain, case_id):
     rebuild a case after archiving / deleting it
     """
 
-    try:
-        case = CommCareCase.get(case_id)
-        found = True
-    except ResourceNotFound:
-        case = CommCareCase()
-        case._id = case_id
-        found = False
-
-    forms = get_case_forms(case_id)
-    filtered_forms = [f for f in forms if f.doc_type == "XFormInstance"]
-    sorted_forms = sorted(filtered_forms, key=lambda f: f.received_on)
-
-    actions, domain = _get_actions_from_forms(sorted_forms, case_id)
-
-    if not found and case.domain is None:
-        case.domain = domain
-
-    rebuild_case_from_actions(case, actions)
-    # todo: should this move to case.rebuild?
-    if not case.xform_ids:
-        if not found:
-            return None
-        # there were no more forms. 'delete' the case
-        case.doc_type = 'CommCareCase-Deleted'
-
-    # add a "rebuild" action
-    case.actions.append(_rebuild_action())
-    case.save()
-    return case
+    return FormProcessorInterface(domain).hard_rebuild_case(case_id)
 
 
 def safe_hard_delete(case):
@@ -141,11 +93,3 @@ def get_case_forms(case_id):
     form_ids = get_case_xform_ids(case_id)
     return [fetch_and_wrap_form(id) for id in form_ids]
 
-
-def _rebuild_action():
-    now = datetime.utcnow()
-    return CommCareCaseAction(
-        action_type=const.CASE_ACTION_REBUILD,
-        date=now,
-        server_date=now,
-    )
