+++ corehq/form_processor/backends/couch/processor.py
@@ -1,6 +1,10 @@
 import datetime
 
-from casexml.apps.case.models import CommCareCase
+from casexml.apps.case import const
+from casexml.apps.case.cleanup import get_case_forms, rebuild_case_from_actions
+from casexml.apps.case.models import CommCareCase, CommCareCaseAction
+from casexml.apps.case.xform import get_case_updates
+from corehq.form_processor.exceptions import CaseNotFound
 from couchforms.util import process_xform, deprecation_type
 from couchforms.models import XFormInstance, XFormDeprecated, XFormDuplicate, doc_types, XFormError
 from corehq.util.couch_helpers import CouchAttachmentsBuilder
@@ -115,3 +119,64 @@ class FormProcessorCouch(object):
     @classmethod
     def xformerror_from_xform_instance(self, instance, error_message, with_new_id=False):
         return XFormError.from_xform_instance(instance, error_message, with_new_id=with_new_id)
+
+    @staticmethod
+    def hard_rebuild_case(case_id):
+        try:
+            case = CommCareCase.get(case_id)
+            found = True
+        except CaseNotFound:
+            case = CommCareCase()
+            case.case_id = case_id
+            found = False
+
+        forms = get_case_forms(case_id)
+        filtered_forms = [f for f in forms if f.doc_type == "XFormInstance"]
+        sorted_forms = sorted(filtered_forms, key=lambda f: f.received_on)
+
+        actions, domain = _get_actions_from_forms(sorted_forms, case_id)
+
+        if not found and case.domain is None:
+            case.domain = domain
+
+        rebuild_case_from_actions(case, actions)
+        # todo: should this move to case.rebuild?
+        if not case.xform_ids:
+            if not found:
+                return None
+            # there were no more forms. 'delete' the case
+            case.doc_type = 'CommCareCase-Deleted'
+
+        # add a "rebuild" action
+        case.actions.append(_rebuild_action())
+        case.save()
+        return case
+
+
+def _get_actions_from_forms(sorted_forms, case_id):
+    from corehq.apps.commtrack.processing import get_stock_actions
+    case_actions = []
+    domain = None
+    for form in sorted_forms:
+        if domain is None:
+            domain = form.domain
+        assert form.domain == domain
+
+        case_updates = get_case_updates(form)
+        filtered_updates = [u for u in case_updates if u.id == case_id]
+        for u in filtered_updates:
+            case_actions.extend(u.get_case_actions(form))
+        stock_actions = get_stock_actions(form)
+        case_actions.extend([intent.action
+                             for intent in stock_actions.case_action_intents
+                             if not intent.is_deprecation])
+    return case_actions, domain
+
+
+def _rebuild_action():
+    now = datetime.utcnow()
+    return CommCareCaseAction(
+        action_type=const.CASE_ACTION_REBUILD,
+        date=now,
+        server_date=now,
+    )
