+++ corehq/form_processor/interfaces/processor.py
@@ -129,3 +129,6 @@ class FormProcessorInterface(object):
 
     def assign_new_id(self, xform):
         return self.processor.assign_new_id(xform)
+
+    def hard_rebuild_case(self, case_id):
+        return self.processor.hard_rebuild_case(case_id)
